stress-revio: remove redundant double cast

+++ stress-revio.c
@@ -472,8 +472,7 @@ PRAGMA_UNROLL_N(4)
 
 	if ((iterations > 0) && (avg_extents > 0.0)) {
 		avg_extents /= (double)iterations;
-		stress_metrics_set(args, "extents",
-			(double)avg_extents, STRESS_METRIC_GEOMETRIC_MEAN);
+		stress_metrics_set(args, "extents", avg_extents, STRESS_METRIC_GEOMETRIC_MEAN);
 	}
 
 	rc = EXIT_SUCCESS;
